cross_mode_retrieval/requirements.txt
torch
torchvision
transformers
open-clip-torch
Pillow
numpy
tqdm
requests
